1、remove issue field of Tag 2、remove IsPublic flag 3、TagClass owner can be zero address 3、combine newTag and update to setTag

+++ contracts/WriteBuffer.sol
@@ -369,7 +369,7 @@ library WriteBuffer {
         return buf.buf;
     }
 
-    function length(buffer memory buf) internal pure returns (uint){
+    function length(buffer memory buf) internal pure returns (uint256) {
         return buf.buf.length;
     }
 }
